+++ Changes.md
@@ -3,6 +3,7 @@
 
 - Revert back to Celluloid's TaskFiber for job processing which has proven to be more
   stable than TaskThread. [#985]
+- Avoid possible lockup during hard shutdown [#997]
 
 At this point, if you are experiencing stability issues with Sidekiq in
 Ruby 1.9, please try Ruby 2.0.  It seems to be more stable.